+++ model/champion_data_info.go
@@ -1,0 +1,9 @@
+package model
+
+// ChampionDataInfo contains information about the playstyle of a champion
+type ChampionDataInfo struct {
+	Attack     int `json:"attack"`
+	Defense    int `json:"defense"`
+	Magic      int `json:"magic"`
+	Difficulty int `json:"difficulty"`
+}
