+++ model/image_data.go
@@ -1,0 +1,12 @@
+package model
+
+// ImageData contains information about an image
+type ImageData struct {
+	Full   string `json:"full"`
+	Sprite string `json:"sprite"`
+	Group  string `json:"group"`
+	X      int    `json:"x"`
+	Y      int    `json:"y"`
+	W      int    `json:"w"`
+	H      int    `json:"h"`
+}
