+++ model/legacy_mastery.go
@@ -1,0 +1,7 @@
+package model
+
+// LegacyMastery represents the old masteries
+type LegacyMastery struct {
+	MasteryID int `json:"masteryId"`
+	Rank      int `json:"rank"`
+}
