+++ model/rune.go
@@ -1,0 +1,7 @@
+package model
+
+// Rune represents an old rune
+type Rune struct {
+	RuneID int `json:"runeId"`
+	Rank   int `json:"rank"`
+}
