+++ model/service.go
@@ -1,0 +1,9 @@
+package model
+
+// Service is a service provided by Riot with its status
+type Service struct {
+	Status    string     `json:"status"`
+	Incidents []Incident `json:"incidents"`
+	Name      string     `json:"name"`
+	Slug      string     `json:"slug"`
+}
