+++ model/status.go
@@ -9,37 +9,3 @@ type Status struct {
 	Slug      string    `json:"slug"`
 	Locales   []string  `json:"locales"`
 }
-
-// Service is a service provided by Riot with its status
-type Service struct {
-	Status    string     `json:"status"`
-	Incidents []Incident `json:"incidents"`
-	Name      string     `json:"name"`
-	Slug      string     `json:"slug"`
-}
-
-// Incident contains information about an incident
-type Incident struct {
-	Active    bool            `json:"active"`
-	CreatedAt string          `json:"created_at"`
-	ID        int             `json:"id"`
-	Updates   []StatusMessage `json:"updates"`
-}
-
-// StatusMessage contains information about a status message
-type StatusMessage struct {
-	Severity     string              `json:"severity"`
-	Author       string              `json:"author"`
-	CreatedAt    string              `json:"created_at"`
-	Translations []StatusTranslation `json:"translations"`
-	UpdatedAt    string              `json:"updated_at"`
-	Content      string              `json:"content"`
-	ID           string              `json:"id"`
-}
-
-// StatusTranslation contains the status message content in a certain language
-type StatusTranslation struct {
-	Locale    string `json:"locale"`
-	Content   string `json:"content"`
-	UpdatedAt string `json:"updated_at"`
-}
