+++ model/team_ban.go
@@ -1,0 +1,9 @@
+package model
+
+// TeamBan is a champion banned by a team
+type TeamBan struct {
+	// Turn during which the champion was banned.
+	PickTurn int `json:"pickTurn"`
+	// Banned championId.
+	ChampionID int `json:"championId"`
+}
